+++ analytics.js
@@ -1,4 +1,4 @@
-//     Analytics.js 0.3.3
+//     Analytics.js 0.3.4
 
 //     (c) 2013 Segment.io Inc.
 //     Analytics.js may be freely distributed under the MIT license.
@@ -381,11 +381,11 @@ analytics.addProvider('Clicky', {
 });
 
 
-// Comscore
+// comScore
 // ---------
 // [Documentation](http://direct.comscore.com/clients/help/FAQ.aspx#faqTagging)
 
-analytics.addProvider('Comscore', {
+analytics.addProvider('comScore', {
 
     settings : {
         c1 : '2',
@@ -1042,14 +1042,14 @@ analytics.addProvider('Mixpanel', {
 
         (function(c,a){window.mixpanel=a;var b,d,h,e;b=c.createElement("script");
         b.type="text/javascript";b.async=!0;b.src=("https:"===c.location.protocol?"https:":"http:")+
-        '//cdn.mxpnl.com/libs/mixpanel-2.1.min.js';d=c.getElementsByTagName("script")[0];
+        '//cdn.mxpnl.com/libs/mixpanel-2.2.min.js';d=c.getElementsByTagName("script")[0];
         d.parentNode.insertBefore(b,d);a._i=[];a.init=function(b,c,f){function d(a,b){
         var c=b.split(".");2==c.length&&(a=a[c[0]],b=c[1]);a[b]=function(){a.push([b].concat(
         Array.prototype.slice.call(arguments,0)))}}var g=a;"undefined"!==typeof f?g=a[f]=[]:
         f="mixpanel";g.people=g.people||[];h=['disable','track','track_pageview','track_links',
-        'track_forms','register','register_once','unregister','identify','name_tag',
-        'set_config','people.identify','people.set','people.increment'];for(e=0;e<h.length;e++)d(g,h[e]);
-        a._i.push([b,c,f])};a.__SV=1.1;})(document,window.mixpanel||[]);
+        'track_forms','register','register_once','unregister','identify','alias','name_tag',
+        'set_config','people.set','people.increment'];for(e=0;e<h.length;e++)d(g,h[e]);
+        a._i.push([b,c,f])};a.__SV=1.2;})(document,window.mixpanel||[]);
 
         // Pass settings directly to `init` as the second argument.
         window.mixpanel.init(this.settings.token, this.settings);
@@ -1082,7 +1082,6 @@ analytics.addProvider('Mixpanel', {
         if (userId) {
             window.mixpanel.identify(userId);
             if (this.settings.nameTag) window.mixpanel.name_tag(userId);
-            if (this.settings.people) window.mixpanel.people.identify(userId);
         }
         if (traits) {
             window.mixpanel.register(traits);