+++ History.md
@@ -1,7 +1,13 @@
+
+0.3.4 / 2013-1-13
+==================
+
+* Upgrade to Mixpanel 2.2 by [@loganfuller](https://github.com/loganfuller)
+
 0.3.3 / 2013-1-11
 ==================
 
-* [Comscore Direct](http://direct.comscore.com) support added
+* [comScore Direct](http://direct.comscore.com) support added
 
 0.3.2 / 2013-1-11
 ==================